+++ EspaperClient/Board.h
@@ -53,7 +53,7 @@ class BoardClass {
 
         String getChipId();
 
-        void deepSleep(uint32_t sleepSeconds);
+        void deepSleep(uint64_t sleepSeconds);
 
         boolean connectWifi(String wifiSsid, String wifiPassword);
